+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/contracts/Contracts.kt
@@ -217,10 +217,14 @@ class Contracts {
 
   data class ConfigureRequest (
     val findTimeoutMillis: Long,
-    val androidDontSuppressAccessibilityServices: Boolean? = null
+    val androidDontSuppressAccessibilityServices: Boolean? = null,
+    val androidRetrieveInteractiveWindows: Boolean? = null
   ){
     fun hasAndroidDontSuppressAccessibilityServices(): Boolean {
       return androidDontSuppressAccessibilityServices != null
+    }
+    fun hasAndroidRetrieveInteractiveWindows(): Boolean {
+      return androidRetrieveInteractiveWindows != null
     }
   }