+++ packages/patrol/darwin/patrol/Sources/PatrolImpl/AutomatorServer/Contracts.swift
@@ -213,6 +213,7 @@ public struct RunDartTestResponse: Codable {
 public struct ConfigureRequest: Codable {
   public var findTimeoutMillis: Int
   public var androidDontSuppressAccessibilityServices: Bool?
+  public var androidRetrieveInteractiveWindows: Bool?
 }
 
 public struct OpenAppRequest: Codable {