+++ packages/patrol_devtools_extension/lib/api/contracts.dart
@@ -292,6 +292,7 @@ class ConfigureRequest with Equatable {
   const ConfigureRequest({
     required this.findTimeoutMillis,
     this.androidDontSuppressAccessibilityServices,
+    this.androidRetrieveInteractiveWindows,
   });
 
   factory ConfigureRequest.fromJson(Map<String, dynamic> json) =>
@@ -299,6 +300,7 @@ class ConfigureRequest with Equatable {
 
   final int findTimeoutMillis;
   final bool? androidDontSuppressAccessibilityServices;
+  final bool? androidRetrieveInteractiveWindows;
 
   Map<String, dynamic> toJson() => _$ConfigureRequestToJson(this);
 
@@ -306,6 +308,7 @@ class ConfigureRequest with Equatable {
   List<Object?> get props => [
     findTimeoutMillis,
     androidDontSuppressAccessibilityServices,
+    androidRetrieveInteractiveWindows,
   ];
 }